requirements.txt
streamlit
pandas
matplotlib.pyplot
uploaded_file_manager